+++ src/Connection.js
@@ -4,27 +4,11 @@ import StateMachine from 'javascript-state-machine';
 import { debug, error } from './Logger';
 import PlayError from './PlayError';
 import PlayErrorCode from './PlayErrorCode';
-import { sdkVersion, protocolVersion } from './Config';
-import { serializeObject } from './CodecUtils';
 import { adapters } from './PlayAdapter';
-
-// eslint-disable-next-line camelcase
-const google_protobuf_wrappers_pb = require('google-protobuf/google/protobuf/wrappers_pb.js');
-
-// eslint-disable-next-line camelcase
-const { BoolValue } = google_protobuf_wrappers_pb;
 
 const protocol = require('./proto/messages_pb');
 
-const {
-  Command,
-  Body,
-  RoomOptions,
-  SessionOpenRequest,
-  RequestMessage,
-  CommandType,
-  OpType,
-} = protocol;
+const { Command, Body, CommandType, OpType } = protocol;
 
 const CommandTypeSwap = Object.keys(CommandType).reduce(
   (obj, key) => Object.assign({}, obj, { [CommandType[key]]: key }),
@@ -36,65 +20,9 @@ const OpTypeSwap = Object.keys(OpType).reduce(
 );
 
 const MAX_NO_PONG_TIMES = 2;
-const MAX_PLAYER_COUNT = 10;
 
 export const ERROR_EVENT = 'ERROR_EVENT';
 export const DISCONNECT_EVENT = 'DISCONNECT_EVENT';
-
-export function convertToRoomOptions(roomName, options, expectedUserIds) {
-  const roomOptions = new RoomOptions();
-  if (roomName) {
-    roomOptions.setCid(roomName);
-  }
-  if (options) {
-    const {
-      open,
-      visible,
-      emptyRoomTtl,
-      playerTtl,
-      maxPlayerCount,
-      customRoomProperties,
-      customRoomPropertyKeysForLobby,
-      flag,
-      pluginName,
-    } = options;
-    if (open !== undefined) {
-      const o = new BoolValue();
-      o.setValue(open);
-      roomOptions.setOpen(open);
-    }
-    if (visible !== undefined) {
-      const v = new BoolValue();
-      v.setValue(visible);
-      roomOptions.setVisible(v);
-    }
-    if (emptyRoomTtl > 0) {
-      roomOptions.setEmptyRoomTtl(emptyRoomTtl);
-    }
-    if (playerTtl > 0) {
-      roomOptions.setPlayerTtl(playerTtl);
-    }
-    if (maxPlayerCount > 0 && maxPlayerCount < MAX_PLAYER_COUNT) {
-      roomOptions.setMaxMembers(maxPlayerCount);
-    }
-    if (customRoomProperties) {
-      roomOptions.setAttr(serializeObject(customRoomProperties));
-    }
-    if (customRoomPropertyKeysForLobby) {
-      roomOptions.setLobbyAttrKeysList(customRoomPropertyKeysForLobby);
-    }
-    if (flag !== undefined) {
-      roomOptions.setFlag(flag);
-    }
-    if (pluginName) {
-      roomOptions.setPluginName(pluginName);
-    }
-  }
-  if (expectedUserIds) {
-    roomOptions.setExpectMembersList(expectedUserIds);
-  }
-  return roomOptions;
-}
 
 /* eslint class-methods-use-this: ["error", { "exceptMethods": ["_getPingDuration", "_handleNotification", "_handleErrorMsg", "_handleUnknownMsg", "_getFastOpenUrl"] }] */
 export default class Connection extends EventEmitter {
@@ -114,7 +42,11 @@ export default class Connection extends EventEmitter {
         { name: 'connected', from: 'connecting', to: 'connected' },
         { name: 'connectFailed', from: 'connecting', to: 'init' },
         { name: 'disconnect', from: 'connected', to: 'disconnected' },
-        { name: 'close', from: 'connected', to: 'init' },
+        {
+          name: 'close',
+          from: ['init', 'connecting', 'connected', 'disconnected'],
+          to: 'closed',
+        },
       ],
     });
   }
@@ -135,6 +67,14 @@ export default class Connection extends EventEmitter {
       this._ws = new WebSocket(url, 'protobuf.1');
       this._ws.onopen = () => {
         debug(`${this._userId} : ${this._flag} connection open`);
+        if (this._fsm.is('closed')) {
+          this._ws.onopen = null;
+          this._ws.onmessage = null;
+          this._ws.onclose = null;
+          this._ws.onerror = null;
+          this._ws.close();
+          return;
+        }
         this._connected();
       };
       this._ws.onclose = () => {
@@ -193,19 +133,6 @@ export default class Connection extends EventEmitter {
     this._fsm.connected();
   }
 
-  async openSession(appId, userId, gameVersion, sessionToken) {
-    const sessionOpen = new SessionOpenRequest();
-    sessionOpen.setAppId(appId);
-    sessionOpen.setPeerId(userId);
-    sessionOpen.setSdkVersion(sdkVersion);
-    sessionOpen.setGameVersion(gameVersion);
-    sessionOpen.setProtocolVersion(protocolVersion);
-    sessionOpen.setSessionToken(sessionToken);
-    const req = new RequestMessage();
-    req.setSessionOpen(sessionOpen);
-    await this.sendRequest(CommandType.SESSION, OpType.OPEN, req);
-  }
-
   _handleCommand(cmd, op, body) {
     if (body.hasResponse()) {
       // 应答
@@ -235,7 +162,7 @@ export default class Connection extends EventEmitter {
     }
   }
 
-  async sendRequest(cmd, op, req) {
+  sendRequest(cmd, op, req) {
     const msgId = this._getMsgId();
     req.setI(msgId);
     const body = new Body();
@@ -249,18 +176,16 @@ export default class Connection extends EventEmitter {
     });
   }
 
-  async sendCommand(cmd, op, body) {
+  sendCommand(cmd, op, body) {
     const command = new Command();
     command.setCmd(cmd);
     command.setOp(op);
     command.setBody(body.serializeBinary());
-    if (debug) {
-      debug(
-        `${this._userId} : ${this._flag} -> ${CommandTypeSwap[cmd]}/${
-          OpTypeSwap[op]
-        }: ${JSON.stringify(body.toObject())}`
-      );
-    }
+    debug(
+      `${this._userId} : ${this._flag} -> ${CommandTypeSwap[cmd]}/${
+        OpTypeSwap[op]
+      }: ${JSON.stringify(body.toObject())}`
+    );
     this._ws.send(command.serializeBinary());
     // ping
     this._ping();
